INC-68 fix: update styles for layouts, menu, header; refactoring
