+++ features/index.ts
@@ -1,1 +1,3 @@
 export * from "./back";
+export * from "./langSelect";
+export * from "./info";
