+++ features/info/index.ts
@@ -1,0 +1,1 @@
+export { Info } from "./Info";
