+++ features/langSelect/index.ts
@@ -1,0 +1,1 @@
+export { LangSelect } from "./LangSelect";
