+++ shared/components/layout/index.ts
@@ -1,1 +1,2 @@
-export * from "./Layout";
+export { Layout } from "./Layout";
+export { getMainLayout } from "./mainLayout";
