+++ shared/components/pageWrapper/pageWrapper.module.scss
@@ -1,8 +1,8 @@
-@import "../../../styles/mixins";
 .root {
-  @include flex(column, center, center);
   position: relative;
-  //width: 100%;
-  min-height: 100vh;
-  margin: 0 auto;
+  margin-left: 220px;
+  @media screen and (max-width: 768px) {
+    margin-left: 0;
+    margin-bottom: 150px;
+  }
 }
