+++ shared/components/select/select.module.scss
@@ -28,8 +28,10 @@
     width: 50px;
     padding: 0 6px;
   }
+
   &.isHiddenText {
-    width: 84px;
+    width: 60px;
+    border: none;
   }
 
   &:hover {
@@ -74,8 +76,9 @@
     width: 48px;
     padding: 0 6px;
   }
+
   &.isHiddenText {
-    width: 82px;
+    width: 58px;
   }
 }
 
@@ -100,4 +103,3 @@
 .label {
   color: var(--color-light-900);
 }
-
