+++ widgets/general/ui/general.module.scss
@@ -1,9 +1,6 @@
 @import "styles/mixins";
 
 .root {
-  @include flex(column, flex-start, center, 24px);
-  margin: 0 30px 0 4px;
-  min-height: 100vh;
   @media screen and (max-width: 576px) {
     margin: 0;
   }
