+++ widgets/index.ts
@@ -11,3 +11,4 @@ export * from "./logOut";
 export * from "./delete-photo-modal";
 export * from "./profile-photo";
 export * from "./mobileMenuMar";
+export * from "./general";
