+++ widgets/mobileMenuMar/MobileMenuBar.module.scss
@@ -8,5 +8,6 @@
   padding: 18px 0;
   height: 60px;
   border-top: 1px solid var(--color-dark-300);
+  background-color: var(--color-dark-700);
   z-index: 100;
 }
