+++ widgets/my-profile/ui/profile.module.scss
@@ -1,4 +1,5 @@
 @import "styles/mixins";
+
 .info_wrapper {
   @include flex(row, center, center);
   width: 100%;
@@ -31,10 +32,6 @@
 
 .posts {
   width: 100%;
-  min-height: 100vh;
   @include flex(row, space-between, center, 12px);
   flex-wrap: wrap;
 }
-
-
-
